Consolidate code for `icmaa-google-tag-manager`
