+++ src/modules/icmaa-google-tag-manager/index.ts
@@ -8,8 +8,7 @@ import { Logger } from '@vue-storefront/core/lib/logger'
 
 import { icmaaGoogleTagManagerModule } from './store'
 import { afterRegistration, isEnabled } from './hooks/afterRegistration'
-
-import { currentStoreView } from '@vue-storefront/core/lib/multistore'
+import afterEach from './hooks/afterEach'
 
 const initGTM = async ({ store, router, appConfig }) => {
   const { id, debug } = appConfig.googleTagManager
@@ -20,17 +19,8 @@ const initGTM = async ({ store, router, appConfig }) => {
     })
 
     router.afterEach((to, from) => {
-      const name = to.meta.gtm || to.name
-      const storeView = currentStoreView()
-      let dataLayer = (window['dataLayer'] = window['dataLayer'] || [])
-      dataLayer.push({
-        'event': 'icmaa-content-view',
-        'content-name': to.fullPath,
-        'content-view-name': name,
-        'store_code': storeView.storeCode,
-        'customerLoggedIn': store.getters['user/isLoggedIn'],
-        'customerEmail': store.getters['user/getUserEmail']
-      })
+      const viewName = to.meta.gtm || to.name
+      afterEach({ store, viewName, path: to.fullPath })
     })
 
     store.dispatch('icmaaGoogleTagManager/enable', true)
@@ -39,7 +29,7 @@ const initGTM = async ({ store, router, appConfig }) => {
   }
 }
 
-export const IcmaaGoogleTagManagerModule: StorefrontModule = async ({store, router, appConfig}) => {
+export const IcmaaGoogleTagManagerModule: StorefrontModule = async ({ store, router, appConfig }) => {
   store.registerModule('icmaaGoogleTagManager', icmaaGoogleTagManagerModule)
 
   initGTM({ appConfig, router, store })
